% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow-package.R
\docType{package}
\name{iaborrow-package}
\alias{iaborrow-package}
\alias{iaborrow}
\title{The 'iaborrow' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
