% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.bin.R
\name{iaborrow.bin}
\alias{iaborrow.bin}
\title{Conducting simulation study of Bayesian adaptive design incorporating data
from historical controls: binary outcome}
\usage{
iaborrow.bin(
  n.CT, n.CC, n.EC, n.CT1, n.CC1,
  out.prob.CT, out.prob.CC, driftOR,
  cov.C, cov.cor.C, cov.effect.C,
  cov.E, cov.cor.E, cov.effect.E,
  chains=2, iter=4000, warmup=floor(iter/2), thin=1,
  a0, alternative="greater", sig.level=0.025, nsim)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment at final analysis
opportunity.}

\item{n.CC}{Number of patients in concurrent control at final analysis
opportunity.}

\item{n.EC}{Number of patients in external control.}

\item{n.CT1}{Number of patients in concurrent treatment at interim analysis
opportunity.}

\item{n.CC1}{Number of patients in concurrent control at interim analysis
opportunity.}

\item{out.prob.CT}{True probability of outcome in concurrent treatment.}

\item{out.prob.CC}{True probability of outcome in concurrent control.}

\item{driftOR}{Odds ratio between external control and concurrent control
for which the bias should be plotted.}

\item{cov.C}{List of covariates in concurrent data. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.cor.C}{Matrix of correlation coefficients for covariates in
concurrent data, specified as Gaussian copula parameter.}

\item{cov.effect.C}{Vector of covariate effects in concurrent data,
specified as odds ratio.}

\item{cov.E}{List of covariates in external data. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.cor.E}{Matrix of correlation coefficients for covariates in
external data, specified as Gaussian copula parameter.}

\item{cov.effect.E}{Vector of covariate effects in external data,
specified as odds ratio.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=2}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=4000}.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{a0}{Parameter of power prior.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.025}.}

\item{nsim}{Number of simulated trials. The default value is \code{nsim=10}.}
}
\value{
\item{w}{Likelihood ratio statistics.}
\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis opportunity.}
\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis opportunity.}
\item{r1}{\code{TRUE} when significant at interim analysis opportunity;
otherwise \code{FALSE}.}
\item{r2}{\code{TRUE} when significant at final analysis opportunity;
otherwise \code{FALSE}.}
}
\description{
Bayesian adaptive design proposed by Psioda (2018) is implemented, where
historical control data is incorporated at interim decision. The binary
outcome is applicable.
}
\examples{
n.CT  <- 100
n.CC  <- 100
n.EC  <- 100
n.CT1 <- 50
n.CC1 <- 50

out.prob.CT <- c(0.23,0.35)
out.prob.CC <- 0.23
driftOR     <- c(0.6,1.0,1.4)

cov.C <- list(list(dist="norm", mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.E <- list(list(dist="norm", mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.E <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.E <- c(0.1,0.1)

a0 <- 0.5

nsim <- 10

iaborrow.bin(
  n.CT=n.CT, n.CC=n.CC, n.EC=n.EC, n.CT1=n.CT1, n.CC1=n.CC1,
  out.prob.CT=out.prob.CT, out.prob.CC=out.prob.CC, driftOR=driftOR,
  cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
  cov.E=cov.E, cov.cor.E=cov.cor.E, cov.effect.E=cov.effect.E,
  a0=a0, nsim=nsim)
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
