% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.cont.R
\name{iaborrow.cont}
\alias{iaborrow.cont}
\title{Simulation study of using historical control at interim analysis occasion}
\usage{
iaborrow.cont(
  n.HC = 200,
  n.C1 = 50,
  n.T1 = 50,
  n.C2 = 100,
  n.T2 = 100,
  out.mean.HC = c(-0.6, -0.4, -0.2, 0, 0.2, 0.4, 0.6),
  out.sd.HC = 1,
  out.mean.C = 0,
  out.sd.C = 1,
  out.mean.T = 0.46,
  out.sd.T = 1,
  cov.mean.HC = c(0, 0),
  cov.sd.HC = c(1, 1),
  cov.cor.HC = cbind(c(1, 0.5), c(0.5, 1)),
  cov.mean.C = c(0, 0),
  cov.sd.C = c(1, 1),
  cov.cor.C = cbind(c(1, 0.5), c(0.5, 1)),
  cov.mean.T = c(0, 0),
  cov.sd.T = c(1, 1),
  cov.cor.T = cbind(c(1, 0.5), c(0.5, 1)),
  a0 = 0.5,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  alternative = "greater",
  sig.level = 0.05,
  w0 = seq(0, 5, 0.1),
  accept.t1e = 0.1,
  accept.pow = 0.7,
  nsim = 10
)
}
\arguments{
\item{n.HC}{Number of patients in external control.}

\item{n.C1}{Number of patients in concurrent control at interim analysis
occasion.}

\item{n.T1}{Number of patients in treatment at interim analysis
occasion.}

\item{n.C2}{Number of patients in concurrent control at final analysis
occasion.}

\item{n.T2}{Number of patients in treatment at final analysis
occasion.}

\item{out.mean.HC}{True mean of outcome in historical control.}

\item{out.sd.HC}{True sd of outcome in historical control.}

\item{out.mean.C}{True mean of outcome in concurrent control.}

\item{out.sd.C}{True sd of outcome in concurrent control.}

\item{out.mean.T}{True mean of outcome in treatment.}

\item{out.sd.T}{True sd of outcome in treatment.}

\item{cov.mean.HC}{True mean of covariates in historical control.}

\item{cov.sd.HC}{True sd of covariates in historical control.}

\item{cov.cor.HC}{True correlation matrix between covariates in historical
control.}

\item{cov.mean.C}{True mean of covariates in concurrent control.}

\item{cov.sd.C}{True sd of covariates in historical control.}

\item{cov.cor.C}{True correlation matrix between covariates in concurrent
control.}

\item{cov.mean.T}{True mean of covariates in treatment.}

\item{cov.sd.T}{True sd of covariates in treatment.}

\item{cov.cor.T}{True correlation matrix between covariates in treatment.}

\item{a0}{Parameter of power prior.}

\item{chains}{Number of Markov chains in MCMC sampling.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling.}

\item{thin}{Period for saving samples in MCMC sampling.}

\item{alternative}{Alternative hypothesis.}

\item{sig.level}{Significance level.}

\item{w0}{Range of stopping criteria.}

\item{accept.t1e}{Acceptable maximum type I error rate.}

\item{accept.pow}{Acceptable minimum power.}

\item{nsim}{Number of simulated trials.}
}
\value{
Summary of simulation study results
}
\description{
Simulation study of using historical control at interim analysis occasion
}
\details{
The \code{iaborrow.cont} is a function generates the operating
characteristics of the Bayesian Optimal Interval design based on toxicity
and efficacy (BOIN-ET) by simulating trials.
}
\examples{
iaborrow.cont()
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
