% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.summary.R
\name{iaborrow.summary}
\alias{iaborrow.summary}
\title{Summarizing simulation study results}
\usage{
iaborrow.summary(
  n.CT, n.CC, n.CT1, n.CC1, ndrift,
  w, p1, p2, r1, r2,
  w0, accept.t1e=0.1, accept.pow=0.7)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment at final analysis
opportunity.}

\item{n.CC}{Number of patients in concurrent control at final analysis
opportunity.}

\item{n.CT1}{Number of patients in concurrent treatment at interim analysis
opportunity.}

\item{n.CC1}{Number of patients in concurrent control at interim analysis
opportunity.}

\item{ndrift}{Number of drift scenarios.}

\item{w}{Likelihood ratio statistics.}

\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis occasion.}

\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis occasion.}

\item{r1}{\code{TRUE} when significant at interim analysis opportunity;
otherwise \code{FALSE}.}

\item{r2}{\code{TRUE} when significant at final analysis opportunity;
otherwise \code{FALSE}.}

\item{w0}{Candidate values of stoppage criteria.}

\item{accept.t1e}{Maximum acceptable type I error rate. The default value is
\code{accept.t1e=0.1}.}

\item{accept.pow}{Minimum acceptable power. The default value is
\code{accept.pow=0.7}.}
}
\description{
Simulation study results are summarized.
}
