% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.t2e.R
\name{iaborrow.t2e}
\alias{iaborrow.t2e}
\title{Conducting simulation study of Bayesian adaptive design incorporating data
from historical controls: time-to-event outcome}
\usage{
iaborrow.t2e <- function(
  n.CT, n.CC, nevent.C, n.EC, nevent.E, nevent.C1, accrual,
  out.mevent.CT, out.mevent.CC, driftHR,
  cov.CT, cov.CC, cov.EC, cormat,
  chains=2, iter=4000, warmup=floor(iter/2), thin=1,
  a0, alternative="greater", sig.level=0.05, nsim=10)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment at final analysis
occasion.}

\item{n.CC}{Number of patients in concurrent control at final analysis
occasion.}

\item{nevent.C}{Number of events in concurrent data at final analysis
occasion.}

\item{n.EC}{Number of patients in external control.}

\item{nevent.E}{Number of events in external control.}

\item{nevent.C1}{Number of events in concurrent data at interim analysis
occasion.}

\item{accrual}{Accrual rate.}

\item{out.mevent.CT}{True median time to event in concurrent treatment.}

\item{out.mevent.CC}{True median time to event in concurrent control.}

\item{driftHR}{HR between external control and concurrent control
for which the bias should be plotted.}

\item{cov.CT}{List of covariates in concurrent treatment. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.CC}{List of covariates in concurrent control. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.EC}{List of covariates in historical control. Distribution and
and its parameters need to be specified for each covariate.}

\item{cormat}{Matrix of correlation between outcome and covariates,
specified as Gaussian copula parameter.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=4}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=2000}.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{a0}{Parameter of power prior.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.05}.}

\item{nsim}{Number of simulated trials. The default value is \code{nsim=100}.}
}
\value{
\item{w}{Likelihood ratio statistics.}
\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis occasion.}
\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis occasion.}
\item{r1}{Reject the null hypothesis at interim analysis occasion.}
\item{r2}{Reject the null hypothesis at final analysis occasion.}
}
\description{
Bayesian adaptive design proposed by Psioda (2018) is implemented, where
historical control data is incorporated at interim decision. The univariate
continuous outcome is only applicable.
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
