% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.reg.conc.R
\name{norm.reg.conc}
\alias{norm.reg.conc}
\title{Stan function for normal regression model using concurrent data}
\usage{
norm.reg.conc()
}
\value{
Compiled stan function for normal regression model using concurrent
data
}
\description{
Stan function for normal regression model using concurrent data is compiled.
}
