% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.reg.ext.R
\name{norm.reg.ext}
\alias{norm.reg.ext}
\title{Stan function for normal regression model using external data}
\usage{
norm.reg.ext()
}
\value{
Compiled stan function for normal regression model using external
data
}
\description{
Stan function for normal regression model using external data is compiled.
}
