% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.reg.hyb.R
\name{norm.reg.hyb}
\alias{norm.reg.hyb}
\title{Stan function for normal regression model using hybrid data}
\usage{
norm.reg.hyb()
}
\value{
Compiled stan function for normal regression model using hybrid
data
}
\description{
Stan function for normal regression model using hybrid data is compiled.
}
