% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.bin.concurrent.R
\name{stan.bin.concurrent}
\alias{stan.bin.concurrent}
\title{Stan function for concurrent data with binary outcome}
\usage{
stan.bin.concurrent()
}
\value{
Compiled stan function for concurrent data with binary outcome
}
\description{
Stan function for logistic regression model using concurrent data is compiled.
}
