% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.bin.external.R
\name{stan.bin.external}
\alias{stan.bin.external}
\title{Stan function for external data with binary outcome}
\usage{
stan.bin.external()
}
\value{
Compiled stan function for external data with binary outcome
}
\description{
Stan function for logistic regression model using external data is compiled.
}
