% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.bin.hybrid.R
\name{stan.bin.hybrid}
\alias{stan.bin.hybrid}
\title{Stan function for hybrid data with binary outcome}
\usage{
stan.bin.hybrid()
}
\value{
Compiled stan function for hybrid data with binary outcome
}
\description{
Stan function for logistic regression model using hybrid data is compiled.
}
