% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.cont.concurrent.R
\name{stan.cont.concurrent}
\alias{stan.cont.concurrent}
\title{Stan function for concurrent data with continuous outcome}
\usage{
stan.cont.concurrent()
}
\value{
Compiled stan function for concurrent data with continuous outcome
}
\description{
Stan function for normal regression model using concurrent data is compiled.
}
