% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.cont.external.R
\name{stan.cont.external}
\alias{stan.cont.external}
\title{Stan function for normal regression model using external data}
\usage{
stan.cont.external()
}
\value{
Compiled stan function for normal regression model using external
data
}
\description{
Stan function for normal regression model using external data is compiled.
}
