% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.cont.hybrid.R
\name{stan.cont.hybrid}
\alias{stan.cont.hybrid}
\title{Stan function for hybrid data with continuous outcome}
\usage{
stan.cont.hybrid()
}
\value{
Compiled stan function for hybrid data with continuous outcome
}
\description{
Stan function for normal regression model using hybrid data is compiled.
}
