% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.bin.R
\name{iaborrow.bin}
\alias{iaborrow.bin}
\title{Simulation study of Bayesian adaptive design incorporating data from
historical controls: binary outcome}
\usage{
iaborrow.bin(
  n.CT, n.CC, n.EC, n.CT1, n.CC1,
  out.prob.CT, out.prob.CC, driftOR,
  cov.CT, cov.CC, cov.EC, cormat,
  chains=2, iter=4000, warmup=floor(iter/2), thin=1,
  a0, alternative="greater", sig.level=0.05, nsim=10)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment at final analysis
occasion.}

\item{n.CC}{Number of patients in concurrent control at final analysis
occasion.}

\item{n.EC}{Number of patients in external control.}

\item{n.CT1}{Number of patients in concurrent treatment at interim analysis
occasion.}

\item{n.CC1}{Number of patients in concurrent control at interim analysis
occasion.}

\item{out.prob.CT}{True probability of outcome in concurrent treatment.}

\item{out.prob.CC}{True probability of outcome in concurrent control.}

\item{driftOR}{OR between external control and concurrent control for which
the bias should be plotted.}

\item{cov.CT}{List of covariates in concurrent treatment. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.CC}{List of covariates in concurrent control. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.EC}{List of covariates in historical control. Distribution and
and its parameters need to be specified for each covariate.}

\item{cormat}{Matrix of correlation between outcome and covariates,
specified as Gaussian copula parameter.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=2}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=4000}.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{a0}{Parameter of power prior.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.05}.}

\item{nsim}{Number of simulated trials. The default value is \code{nsim=10}.}
}
\value{
\item{w}{Likelihood ratio statistics.}
\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis occasion.}
\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis occasion.}
\item{r1}{Reject the null hypothesis at interim analysis occasion.}
\item{r2}{Reject the null hypothesis at final analysis occasion.}
}
\description{
Bayesian adaptive design proposed by Psioda (2018) is implemented, where
historical control data is incorporated at interim decision. The binary
outcome is applicable.
}
\examples{
n.CT  <- 180
n.CC  <- 180
n.EC  <- 200
n.CT1 <- 90
n.CC1 <- 90

out.prob.CT <- c(0.23,0.35)
out.prob.CC <- 0.23
driftOR     <- c(0.6,1.0,1.4)

cov.CT <- list(list(dist="norm", mean=0,sd=1),
               list(dist="binom",prob=0.2),
               list(dist="binom",prob=0.4))

cov.CC <- list(list(dist="norm", mean=0,sd=1),
               list(dist="binom",prob=0.2),
               list(dist="binom",prob=0.4))

cov.EC <- list(list(dist="norm", mean=0,sd=1),
               list(dist="binom",prob=0.2),
               list(dist="binom",prob=0.4))

cormat <- rbind(c(1.0,0.1,0.2,0.3),
                c(0.1,1.0,0.4,0.5),
                c(0.2,0.4,1.0,0.6),
                c(0.3,0.5,0.6,1.0))

a0 <- 0.5

iaborrow.bin(
  n.CT=n.CT, n.CC=n.CC, n.EC=n.EC, n.CT1=n.CT1, n.CC1=n.CC1,
  out.prob.CT=out.prob.CT, out.prob.CC=out.prob.CC, driftOR=driftOR,
  cov.CT=cov.CT, cov.CC=cov.CC, cov.EC=cov.EC, cormat=cormat,
  a0=a0)
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
