% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.cont.R
\name{iaborrow.cont}
\alias{iaborrow.cont}
\title{Simulation study of Bayesian adaptive design incorporating data from
historical controls: univariate continuous outcome}
\usage{
iaborrow.cont(
  n.HC, n.C1, n.T1, n.C2, n.T2,
  out.mean.HC, out.sd.HC, out.mean.C, out.sd.C, out.mean.T, out.sd.T,
  cov.mean.HC, cov.sd.HC, cov.cor.HC,
  cov.mean.C, cov.sd.C, cov.cor.C, cov.mean.T, cov.sd.T, cov.cor.T,
  a0, chains=4, iter=2000, warmup=floor(iter/2), thin=1,
  alternative="greater", sig.level=0.05,
  w0, accept.t1e=0.1, accept.pow=0.7, nsim=100)
}
\arguments{
\item{n.HC}{Number of patients in external control.}

\item{n.C1}{Number of patients in concurrent control at interim analysis
occasion.}

\item{n.T1}{Number of patients in treatment at interim analysis
occasion.}

\item{n.C2}{Number of patients in concurrent control at final analysis
occasion.}

\item{n.T2}{Number of patients in treatment at final analysis
occasion.}

\item{out.mean.HC}{True mean of outcome in historical control.}

\item{out.sd.HC}{True sd of outcome in historical control.}

\item{out.mean.C}{True mean of outcome in concurrent control.}

\item{out.sd.C}{True sd of outcome in concurrent control.}

\item{out.mean.T}{True mean of outcome in treatment.}

\item{out.sd.T}{True sd of outcome in treatment.}

\item{cov.mean.HC}{Vector of true mean of covariates in historical control.}

\item{cov.sd.HC}{Vector of true sd of covariates in historical control.}

\item{cov.cor.HC}{True correlation matrix between covariates in historical
control.}

\item{cov.mean.C}{Vector of true mean of covariates in concurrent control.}

\item{cov.sd.C}{Vector of true sd of covariates in historical control.}

\item{cov.cor.C}{True correlation matrix between covariates in concurrent
control.}

\item{cov.mean.T}{Vector of true mean of covariates in treatment.}

\item{cov.sd.T}{Vector of true sd of covariates in treatment.}

\item{cov.cor.T}{True correlation matrix between covariates in treatment.}

\item{a0}{Parameter of power prior.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=4}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=2000}.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.05}.}

\item{w0}{Vector of range of stopping criteria.}

\item{accept.t1e}{Acceptable maximum type I error rate. The default value is
\code{accept.t1e=0.1}.}

\item{accept.pow}{Acceptable minimum power. The default value is
\code{accept.pow=0.7}.}

\item{nsim}{Number of simulated trials. The default value is \code{nsim=100}.}
}
\value{
\item{w}{Likelihood ratio statistics.}
\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis occasion.}
\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis occasion.}
\item{sig.rate}{Reject rate.}
\item{max.t1e}{Type I error at every \code{w0}}
\item{min.pow}{Power at every \code{w0}}
\item{stop.rate}{Stoppage probability at every \code{w0}.}
\item{w0.opt}{Optimal value of \code{w0}.}
\item{t1e}{Type I error rate at optimal value of \code{w0}.}
\item{stop.null}{Stoppage probability at optimal value of \code{w0} under
null hypothesis.}
\item{exp.size.null}{Expected total sample size under null hypothesis.}
\item{pow}{Power at optimal value of \code{w0}.}
\item{stop.alt}{Stoppage probability at optimal value of \code{w0} under
alternative hypothesis.}
\item{exp.size.alt}{Expected total sample size under alternative hypothesis.}
}
\description{
Bayesian adaptive design proposed by Psioda (2018) is implemented, where
historical control data is incorporated at interim decision. The univariate
continuous outcome is only applicable.
}
\details{
The \code{iaborrow.cont} is a function which generates the operating
characteristics of Bayesian adaptive design incorporating data from
historical controls. This fixed-borrowing adaptive design incorporates
subject-level control data from a previously completed clinical trial. The
interim analysis evaluates the extent of prior-data conflict based on a
similarity measure for new and historical trial data, and then (i) if the
conflict is too great, the new trial is continued and the data are analyzed
with non-informative prior, (ii) Otherwise, the new trial is stopped early
and the data are analyzed with informative prior. The univariate
continuous outcome is only applicable.
}
\examples{
n.HC        <- 200
n.C1        <- 50
n.T1        <- 50
n.C2        <- 100
n.T2        <- 100
out.mean.HC <- c(-0.6,-0.4,-0.2,0.0,0.2,0.4,0.6)
out.sd.HC   <- 1
out.mean.C  <- 0
out.sd.C    <- 1
out.mean.T  <- 0.46
out.sd.T    <- 1
cov.mean.HC <- c(0,0)
cov.sd.HC   <- c(1,1)
cov.cor.HC  <- cbind(c(1.0,0.5),c(0.5,1.0))
cov.mean.C  <- c(0,0)
cov.sd.C    <- c(1,1)
cov.cor.C   <- cbind(c(1.0,0.5),c(0.5,1.0))
cov.mean.T  <- c(0,0)
cov.sd.T    <- c(1,1)
cov.cor.T   <- cbind(c(1.0,0.5),c(0.5,1.0))
a0          <- 0.5
w0          <- seq(0,5,0.1)

iaborrow.cont(
  n.HC=n.HC, n.C1=n.C1, n.T1=n.T1, n.C2=n.C2, n.T2=n.T2,
  out.mean.HC=out.mean.HC, out.sd.HC=out.sd.HC,
  out.mean.C=out.mean.C, out.sd.C=out.sd.C,
  out.mean.T=out.mean.T, out.sd.T=out.sd.T,
  cov.mean.HC=cov.mean.HC, cov.sd.HC=cov.sd.HC, cov.cor.HC=cov.cor.HC,
  cov.mean.C=cov.mean.C, cov.sd.C=cov.sd.C, cov.cor.C=cov.cor.C,
  cov.mean.T=cov.mean.T, cov.sd.T=cov.sd.T, cov.cor.T=cov.cor.T,
  a0=a0, w0=w0)
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
