% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaborrow.summary.R
\name{iaborrow.summary}
\alias{iaborrow.summary}
\title{Simulation study of Bayesian adaptive design incorporating data from
historical controls: binary outcome}
\usage{
iaborrow.summary(
  n.CT, n.CC, n.CT1, n.CC1, ndrift,
  w, p1, p2, r1, r2,
  w0, accept.t1e=0.1, accept.pow=0.7)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment at final analysis
occasion.}

\item{n.CC}{Number of patients in concurrent control at final analysis
occasion.}

\item{n.CT1}{Number of patients in concurrent treatment at interim analysis
occasion.}

\item{n.CC1}{Number of patients in concurrent control at interim analysis
occasion.}

\item{ndrift}{Number of drift scenarios.}

\item{w}{Likelihood ratio statistics.}

\item{p1}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at interim analysis occasion.}

\item{p2}{Probability meeting that the treatment effect is above 0
(if \code{alternative="greater"}) or below 0 (if \code{alternative="less"})
at final analysis occasion.}

\item{r1}{Reject the null hypothesis at interim analysis occasion.}

\item{r2}{Reject the null hypothesis at final analysis occasion.}
}
\value{
\item{sig.rate}{Reject rate.}
\item{max.t1e}{Type I error at every \code{w0}}
\item{min.pow}{Power at every \code{w0}}
\item{stop.rate}{Stoppage probability at every \code{w0}.}
\item{w0.opt}{Optimal value of \code{w0}.}
\item{t1e}{Type I error rate at optimal value of \code{w0}.}
\item{stop.null}{Stoppage probability at optimal value of \code{w0} under
null hypothesis.}
\item{exp.size.null}{Expected total sample size under null hypothesis.}
\item{pow}{Power at optimal value of \code{w0}.}
\item{stop.alt}{Stoppage probability at optimal value of \code{w0} under
alternative hypothesis.}
\item{exp.size.alt}{Expected total sample size under alternative hypothesis.}
}
\description{
Bayesian adaptive design proposed by Psioda (2018) is implemented, where
historical control data is incorporated at interim decision. The binary
outcome is applicable.
}
\references{
Psioda MA, Soukup M, Ibrahim JG. A practical Bayesian adaptive design
incorporating data from historical controls \emph{Statistics in Medicine}
2018; 37:4054-4070.
}
